requirements.txt
numpy==1.21.4
opencv_python==4.5.4.60
Pillow==9.4.0
